separated components into folders with individual css
