+++ src/App.css
@@ -264,8 +264,7 @@
   width: 100%;
   margin-bottom: 8px;
 }
-.id {
-}
+
 .stateLabel {
   font-weight: bold;
   margin-right: 12px;
