+++ src/components/Button/Button.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import './Button.css';
 
 export default function Button({ text, onClick, type }) {
 	return (
