+++ src/components/CourseInfo/CourseInfo.css
@@ -1,0 +1,37 @@
+.course-info-bottom {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+}
+
+.course-info-wrapper {
+  border: 2px solid teal;
+  padding: 25px;
+}
+.course-info-title {
+  width: 100%;
+  text-align: center;
+}
+.course-info-description {
+  width: 65%;
+}
+
+.stateItem {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  margin-bottom: 8px;
+}
+
+.stateLabel {
+  font-weight: bold;
+  margin-right: 12px;
+}
+.course-info-author {
+  margin-bottom: 8px;
+}
+.id-authors-divs {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+}
