+++ src/components/Courses/Courses.css
@@ -1,0 +1,6 @@
+.courses-wrapper {
+  padding: 25px 25px;
+  display: flex;
+  flex-direction: column;
+  border: 2px solid lightcoral;
+}
