+++ src/components/CreateCourse/CreateCourse.css
@@ -1,0 +1,72 @@
+.create-course-wrapper {
+  border: 2px solid magenta;
+  padding: 25px;
+}
+
+.create-course-wrapper,
+.create-course-top {
+  display: flex;
+  flex-direction: column;
+}
+
+.cc-bottom-left,
+.cc-bottom-right {
+  width: 50%;
+  display: flex;
+  flex-direction: column;
+}
+
+.create-course-bottom {
+  margin-top: 25px;
+  border: 2px solid #902895;
+  padding: 25px;
+}
+
+.create-course-bottom,
+.cc-add-author {
+  display: flex;
+  flex-direction: row;
+}
+
+.cc-add-author > button {
+  margin-top: 55px;
+}
+
+.cc-add-author,
+.cc-add-duration {
+  align-content: flex-start;
+  justify-items: flex-start;
+  justify-content: flex-start;
+  height: 20vh;
+}
+
+.create-author-wrapper {
+  margin-right: 50px;
+}
+
+.create-course-title-entry-wrapper {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.create-course-title-button > input {
+  margin-top: 25px;
+}
+
+.create-course-description-wrapper > textarea {
+  width: 100%;
+  height: 12vh;
+  resize: none;
+}
+
+.cc-bottom-left {
+  align-items: flex-start;
+  align-content: flex-start;
+}
+
+.author-w-button {
+  display: flex;
+  flex-direction: row;
+  margin-bottom: 12px;
+}
