+++ src/components/Header/Header.css
@@ -1,0 +1,20 @@
+.header {
+  border: 2px solid purple;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  padding: 0 25px;
+  vertical-align: middle;
+  margin-bottom: 25px;
+}
+
+.header-text-wrapper {
+  display: flex;
+  flex-direction: row;
+  vertical-align: middle;
+  align-items: center;
+}
+
+.user-info-header {
+  padding-right: 25px;
+}
