+++ src/components/Header/Header.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Button from './Button';
-import Logo from './Logo';
+import Button from '../Button/Button';
+import Logo from '../Logo/Logo';
+import './Header.css';
 
 export default function Header() {
 	return (
