+++ src/components/Login/Login.jsx
@@ -1,6 +1,7 @@
 import { React, useState } from 'react';
 import { Link } from 'react-router-dom';
-import Button from './Button';
+import Button from '../Button/Button';
+import './Login.css';
 
 export default function Login() {
 	const [loginState, setLoginState] = useState({
