+++ src/components/Search/Search.css
@@ -1,0 +1,17 @@
+.search-wrapper {
+  margin-bottom: 25px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+.search-wrapper-form {
+  width: 90%;
+}
+
+.search-course {
+  padding: 5px;
+  margin-right: 15px;
+  height: 20px;
+  min-width: 22%;
+  border-color: cornflowerblue;
+}
